Enable switching to another fov (file)
When the fov combobox changes activation, a slot is triggered to update the
current_fov attr to the new fov and reset the current_molecule to 1. Then the
listview is updated, and a signal (trace_model_should_change_file) is send to
the TraceModel object, which reloads everything with set_data_storage(). Finally
the chartview is updated.

Issue: If the required data file is not found, there will be error that is not
handled, so the view will not update. Change back to valid fov can make view
work again.