Include left censored data in first binding analysis
The first binding analysis is used to analysis the time between flowing
reagent and the first binding event. However there is a dead time
between the injection of reagent and the start point of the recording.
Therefore, there are DNAs that are bound at the beginning, which are
left censored data because the binding time is less than the dead time.
Ignoring these data points will overestimate the binding time.

Changes:

Initially, all the time measurements starts from the recording start
point. The molecules bound at the beginning are marked as
"left_censored", and their durations were set as 0. Then, the dead time
value "time_offset" is added to the "duration" variable to obtain the
true time we want to observe. Those censored point will automatically
have 0 + time_offset upper limit. Note that the plotting tool is changed
to plot_survival_function specifically because left censoring fitting
use cumulative density by default.

Issues:
This analysis accounts for the dead time and the molecules that binds
before acquisition, but cannot include the right censored molecules
(molecules that never binds in the acquisition) because of the limit of
the Kaplan-Meier estimator.