Add test for categorize_binding_traces_script
Test the output against the calculated json data on 2020.02.28.

Some switches and return statement was added in the script to support testing.

Since the loaded json file does not retain the variable type, the result will be
cast before assertion. This will be a temporary test that allows refactoring
and updating packages while preserves the behavior.

The whole dataset of that field of view is included in the test_data directory.
They might be removed if better tests are written.