Fix bug that require manually define channels
Previously the channel definition is hard coded as target: blue and binder:
green. This is not flexible if any other channel combination is used, eg. green
DNA.

I am using the fact that the channelsInfo output from matlab came from the
parameter file, which mean it is ordered and the first channel is the target
channel. The channels_info variable is changed from dict to OrderedDict to
preserve this order. Then the two data attributes are written by a function
define_channels.